Rename agent3 package name to agent
Also brought back the "internal" sub-package.
