+++ agent/src/main/java/com/microsoft/applicationinsights/agent/internal/model/TwoPartCompletion.java
@@ -1,4 +1,4 @@
-package com.microsoft.applicationinsights.agent3.model;
+package com.microsoft.applicationinsights.agent.internal.model;
 
 public class TwoPartCompletion {
